move _iterative_prune and _find_tips to _helpers
rearrage ome imports and change where things are getting called from
