+++ plantcv/plantcv/morphology/__init__.py
@@ -1,6 +1,5 @@
 from plantcv.plantcv.morphology.find_branch_pts import find_branch_pts
-from plantcv.plantcv.morphology.find_tips import _find_tips, find_tips
-from plantcv.plantcv.morphology._iterative_prune import _iterative_prune
+from plantcv.plantcv.morphology.find_tips import find_tips
 from plantcv.plantcv.morphology.segment_skeleton import segment_skeleton
 from plantcv.plantcv.morphology.segment_sort import segment_sort
 from plantcv.plantcv.morphology.prune import prune
@@ -18,8 +17,8 @@ from plantcv.plantcv.morphology.analyze_stem import analyze_stem
 from plantcv.plantcv.morphology.fill_segments import fill_segments
 from plantcv.plantcv.morphology.segment_ends import segment_ends
 
-__all__ = ["find_branch_pts", "_find_tips", "find_tips", "prune", "skeletonize", "check_cycles",
+__all__ = ["find_branch_pts", "find_tips", "prune", "skeletonize", "check_cycles",
            "segment_skeleton", "segment_angle",
            "segment_path_length", "segment_euclidean_length", "segment_curvature", "segment_sort", "segment_id",
-           "segment_tangent_angle", "segment_insertion_angle", "segment_combine", "_iterative_prune", "analyze_stem",
+           "segment_tangent_angle", "segment_insertion_angle", "segment_combine", "analyze_stem",
            "fill_segments", "segment_ends"]
